backend/requirements.txt
amqp==5.3.1
asgiref==3.10.0
async-timeout==5.0.1
attrs==25.4.0
babel==2.17.0
billiard==4.2.2
celery==5.5.3
certifi==2025.10.5
channels==4.3.1
channels_redis==4.3.0
charset-normalizer==3.4.4
click==8.3.0
click-didyoumean==0.3.1
click-plugins==1.1.1.2
click-repl==0.3.0
dj-database-url==3.0.1
Django==5.2.7
django-cors-headers==4.9.0
django-filter==25.2
django-money==3.5.4
djangorestframework==3.16.1
djangorestframework_simplejwt==5.5.1
drf-spectacular==0.28.0
idna==3.11
inflection==0.5.1
jsonschema==4.25.1
jsonschema-specifications==2025.9.1
kombu==5.5.4
msgpack==1.1.2
packaging==25.0
pillow==12.0.0
prompt_toolkit==3.0.52
psycopg2-binary==2.9.11
py-moneyed==3.0
PyJWT==2.10.1
python-dateutil==2.9.0.post0
python-dotenv==1.2.1
PyYAML==6.0.3
redis==7.0.1
referencing==0.37.0
requests==2.32.5
rpds-py==0.28.0
six==1.17.0
sqlparse==0.5.3
typing_extensions==4.15.0
tzdata==2025.2
uritemplate==4.2.0
urllib3==2.5.0
vine==5.1.0
wcwidth==0.2.14
whitenoise==6.11.0
